ERT-1059 Moved RunDialog, SimulationsTracker and changed the iterations model
