+++ devel/python/python/ert_gui/models/connectors/run/CMakeLists.txt
@@ -4,8 +4,6 @@ set(PYTHON_SOURCES
     analysis_module_model.py
     analysis_module_variables_model.py
     iterated_analysis_module_model.py
-    number_of_iterations.py
-    simulations_tracker.py
     target_case_format_model.py
     target_case_model.py
     workflows_model.py
