+++ devel/python/python/ert_gui/pages/CMakeLists.txt
@@ -2,7 +2,6 @@ set(PYTHON_SOURCES
     __init__.py
     ert_summary.py
     popup_dialog.py
-    run_dialog.py
     summary_panel.py
 )
