+++ devel/python/python/ert_gui/simulation/models/CMakeLists.txt
@@ -5,6 +5,7 @@ set(PYTHON_SOURCES
     ensemble_smoother.py
     iterated_ensemble_smoother.py
     multiple_data_assimilation.py
+    simulations_tracker.py
 )
 
 add_python_package("python.ert_gui.simulation.models" ${PYTHON_INSTALL_PREFIX}/ert_gui/simulation/models "${PYTHON_SOURCES}" True)
