+++ devel/python/python/ert_gui/simulation/models/__init__.py
@@ -3,3 +3,4 @@ from .ensemble_experiment import EnsembleExperiment
 from .ensemble_smoother import EnsembleSmoother
 from .iterated_ensemble_smoother import IteratedEnsembleSmoother
 from .multiple_data_assimilation import MultipleDataAssimilation
+from .simulations_tracker import SimulationsTracker
